pyproject.toml
[project]
name = "realtimeaivoicechat"
version = "0.1.0"
description = "A real-time AI voice chat application using FastAPI and LangChain."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "faster-whisper>=1.2.0",
    "langchain-google-genai>=2.1.10",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.31",
    "langgraph>=0.6.7",
    "loguru>=0.7.3",
    "mcp-use>=1.3.10",
    "noisereduce>=3.0.3",
    "numpy>=2.3.2",
    "openai[voice-helpers]>=1.99.9",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.13.1",
    "scipy>=1.16.1",
    "silero-vad>=5.1.2",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "websockets>=15.0.1",
]
